ipacm: Changes to handle calculation overflow
Added changes to handle IPACM file size calculation overflow while
calculating the max file size.

(cherry picked from commit 9c4a769caed2f9ea13e7af20583ec9448f18bdc3)

+++ ipacm/inc/IPACM_Xml.h
@@ -363,7 +363,7 @@ typedef struct  _IPACM_conf_t
 #ifdef FEATURE_IPACM_PER_CLIENT_STATS
 	bool lan_stats_enable;
 #endif
-	uint32_t max_file_size;
+	int64_t max_file_size;
 	uint8_t max_file_size_quota;
 	bool ipv6_nat_enable;
 	int ipacm_l2tp_enable;
